transports/websocket: rename websocket.js to ws.js
