+++ src/transports/index.js
@@ -1,4 +1,4 @@
 export * from './tcp';
 export * from './udp';
 export * from './tls';
-export * from './websocket';
+export * from './ws';
